Turn PS/1 Audio channel wake and sleep routine into functions
Moves the sleep rountine to the end of the channel to ensure
it gets the requested number of bytes before being disabled.

Also drops the DAC reset when we sleep the channel.  The DACs
state will be in a stead-state after 5 seconds without any
reads or writes.

Likewise, DOSBox's audio channel (whether it's playing or not)
is not information the DAC needs to be aware of or have its
state maniplated.  Testing also shows no tangible change.